requirements.txt
#Django
Django==2.1.7
pytz==2018.9
#Postgresql
psycopg2-binary==2.7.7
#Django Rest Framework
djangorestframework==3.10.3